requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1
supabase==2.11.0
openai==1.58.1
apify-client==1.8.1
feedparser==6.0.11
google-api-python-client==2.157.0
resend==2.21.0
jinja2==3.1.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
